+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `val(i,a,b)` now returns the same value like Filter Factory for illegal values of `i`.
 - Windows: Standalone filters containing a `&` in Name or Category now have the correct "Visual Themes", as the Manifest XML is not broken anymore.
 - Introduced obfuscation version 7, which is more secure.
+- Windows: Handling of ampersand characters ("&") in Title, Category, Author, Copyright, and Controller/Map labels fixed.
 
 (*) This bug/solution was tested on Windows but needs to be verified and/or implemented on Mac.